+++ fili-wikipedia-example/src/main/java/com/yahoo/wiki/webservice/data/config/table/WikiTableLoader.java
@@ -109,7 +109,12 @@ public class WikiTableLoader extends BaseTableLoader {
                     dictionaries
             );
             Set<Granularity> validGranularities = validGrains.get(table);
-            loadLogicalTableWithGranularities(LogicalTableName.forName(table), tableGroup, validGranularities, dictionaries);
+            loadLogicalTableWithGranularities(
+                    LogicalTableName.forName(table),
+                    tableGroup,
+                    validGranularities,
+                    dictionaries
+            );
         }
     }
 }